# app/swagger/swagger.yaml
basePath: /
definitions:
...
  request.CabinInReceive:
    properties:
      domain:
        type: string
      method:
        type: string
      pType:
        type: string
      resource:
        description: 路径
        type: string
      source:
        type: string
    type: object
  request.CopyRoleParams:
    properties:
      copyId:
        type: string
      created_at:
        type: string
      domain:
        description: 域
        type: integer
      enable:
        description: 是否启用//radio/1,启用,2,禁用
        type: integer
      id:
        description: 主键
        type: integer
      name:
        description: 名称
        type: string
      parentId:
        description: 角色
        type: string
      remark:
        description: 说明//textarea
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态//radio/2,隐藏,1,显示
        type: integer
      updated_at:
        type: string
    type: object
  request.DataDelete:
    properties:
      ids:
        type: string
    type: object
  request.PageRequest:
    properties:
      filter:
        items:
          type: string
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
...
  response.PageResult:
    properties:
      columns: {}
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  role.DevopsSysRole:
    properties:
      created_at:
        type: string
      domain:
        description: 域
        type: integer
      enable:
        description: 是否启用//radio/1,启用,2,禁用
        type: integer
      id:
        description: 主键
        type: integer
      name:
        description: 名称
        type: string
      parentId:
        description: 角色
        type: string
      remark:
        description: 说明//textarea
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态//radio/2,隐藏,1,显示
        type: integer
      updated_at:
        type: string
    type: object
  sys.ChangePasswordRequest:
    properties:
      oldPassword:
        type: string
      password:
        type: string
      type:
        type: integer
      username:
        type: string
    type: object
...
  sys.LoginRequest:
    properties:
      password:
        type: string
      type:
        type: integer
      username:
        type: string
    type: object
...
  third.AddMerchantApmRequest:
    properties:
      merchant_id:
        description: 商户id
        type: string
      merchant_name:
        description: 商户名称
        type: string
      namespaces:
        description: 商户环境
        items:
          type: string
        type: array
    type: object
...
  third.TencentResourceListRequest:
    properties:
      ids:
        items:
          type: string
        type: array
      limit:
        type: integer
      names:
        items:
          type: string
        type: array
      offset:
        type: integer
      type:
        type: integer
    type: object
  user.DevopsSysUserEntity:
    properties:
      address:
        description: 地址
        type: string
      createID:
        description: 创建者
        type: integer
      created_at:
        type: string
      email:
        description: email
        type: string
      enable:
        description: 是否启用//radio/1,启用,2,禁用
        type: integer
      endTime:
        description: 结束时间
        type: string
      id:
        description: 主键
        type: integer
      merchants:
        description: 所属商户
        type: string
      password:
        description: 密码
        type: string
      realName:
        description: 真实姓名
        type: string
      remark:
        description: 说明
        type: string
      roleIds:
        items:
          type: string
        type: array
      salt:
        description: 密码盐
        type: string
      status:
        description: 状态
        type: integer
      tel:
        description: 手机号
        type: string
      theme:
        description: 主题
        type: string
      titleURL:
        description: 头像地址
        type: string
      updateID:
        description: 更新人
        type: integer
      updated_at:
        type: string
      userType:
        description: 用户类型
        type: integer
      username:
        description: 登录名/11111
        type: string
      uuid:
        description: UUID
        type: string
      workNum:
        description: 工号
        type: string
    type: object
  user.DevopsSysUserView:
    properties:
      address:
        description: 地址
        type: string
      createID:
        description: 创建者
        type: integer
      created_at:
        type: string
      email:
        description: email
        type: string
      enable:
        description: 是否启用//radio/1,启用,2,禁用
        type: integer
      endTime:
        description: 结束时间
        type: string
      id:
        description: 主键
        type: integer
      merchants:
        description: 所属商户
        type: string
      password:
        description: 密码
        type: string
      realName:
        description: 真实姓名
        type: string
      remark:
        description: 说明
        type: string
      roleIds:
        items:
          type: string
        type: array
      roleName:
        type: string
      salt:
        description: 密码盐
        type: string
      status:
        description: 状态
        type: integer
      tel:
        description: 手机号
        type: string
      theme:
        description: 主题
        type: string
      titleURL:
        description: 头像地址
        type: string
      updateID:
        description: 更新人
        type: integer
      updated_at:
        type: string
      userType:
        description: 用户类型
        type: integer
      username:
        description: 登录名/11111
        type: string
      uuid:
        description: UUID
        type: string
      workNum:
        description: 工号
        type: string
    type: object
...
info:
  contact: {}
  description: This is Devops API
  title: Swagger Devops API
  version: 0.0.1
paths:
...
  /sys/roles/{id}:
    get:
      consumes:
      - application/json
      description: 获得角色接口
      parameters:
      - description: 角色id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得角色接口
      tags:
      - Role
  /sys/roles/add:
    post:
      consumes:
      - application/json
      description: 新增角色接口
      parameters:
      - description: 角色
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/role.DevopsSysRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 新增角色接口
      tags:
      - Role
  /sys/roles/add/resources:
    post:
      consumes:
      - application/json
      description: 给角色新增权限接口
      parameters:
      - description: 新增api权限Ptype为p； 新增菜单权限Ptype为p3 , source 是角色的id，resource 是资源
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/request.CabinInReceive'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 给角色新增权限接口
      tags:
      - Role
  /sys/roles/copy:
    post:
      consumes:
      - application/json
      description: 复制角色接口
      parameters:
      - description: 'copyId '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CopyRoleParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 复制角色接口
      tags:
      - Role
  /sys/roles/delete:
    delete:
      consumes:
      - application/json
      description: 删除角色接口
      parameters:
      - description: 角色ids
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/request.DataDelete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除角色接口
      tags:
      - Role
  /sys/roles/list:
    post:
      consumes:
      - application/json
      description: 获得角色列表接口
      parameters:
      - description: 页数，页大小，筛选条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得角色列表接口
      tags:
      - Role
  /sys/roles/modify:
    post:
      consumes:
      - application/json
      description: 修改角色接口
      parameters:
      - description: 角色
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/role.DevopsSysRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改角色接口
      tags:
      - Role
  /third/tencent/apm/addMerchantApm:
    post:
      consumes:
      - application/json
      description: 新增商户apm
      parameters:
      - description: 商户名称，商户id，名称空间
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/third.AddMerchantApmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 新增商户apm
      tags:
      - ThirdTencent
...
  /third/tencent/resource/list:
    post:
      consumes:
      - application/json
      description: 获得腾讯云资源列表接口
      parameters:
      - description: 页数，页大小，筛选条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/third.TencentResourceListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得腾讯云资源列表接口
      tags:
      - ThirdTencent
  /user/add:
    post:
      consumes:
      - application/json
      description: 用户新增接口
      parameters:
      - description: 用户的相关字段， 用户的旧密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.DevopsSysUserEntity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户新增接口
      tags:
      - User
  /user/changePassword:
    post:
      consumes:
      - application/json
      description: 用户修改密码接口
      parameters:
      - description: 用户名，原密码， 新密码，账户类型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sys.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户修改密码接口
      tags:
      - User
  /user/delete:
    delete:
      consumes:
      - application/json
      description: 用户删除接口
      parameters:
      - description: 传用户id, 多个以, 隔开
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/request.DataDelete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户删除接口
      tags:
      - User
  /user/info:
    get:
      description: 根据token获取id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.DevopsSysUserView'
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情的接口
      tags:
      - User
  /user/list:
    post:
      description: 根据参数获取用户列表
      parameters:
      - description: 分页查询
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PageResult'
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表的接口
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 用户名，密码，账户类型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sys.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户登录接口
      tags:
      - User
  /user/logout:
    post:
      description: 用户退出接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户退出接口
      tags:
      - User
  /user/modify:
    put:
      consumes:
      - application/json
      description: 用户修改接口
      parameters:
      - description: 用户的相关字段， 用户的旧密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.DevopsSysUserEntity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户修改接口
      tags:
      - User
  /user/register:
    post:
      description: 用户注册接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户注册接口
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: bearer
    type: apiKey
swagger: "2.0"